requirements.txt
Pillow >= 10.3.0
